requirements.txt
pop==13
acct==2.0
idem==7.1
pop-config==6
takara==1.2
dict-toolbox==1.5
azure-common==1.1.23
azure-core==1.3.0
azure-graphrbac==0.40.0
azure-identity==1.3.0
azure-keyvault-keys==4.0.0
azure-keyvault-secrets==4.1.0
azure-mgmt-applicationinsights==0.2.0
azure-mgmt-authorization==0.50.0
azure-mgmt-compute==4.6.2
azure-mgmt-containerregistry==3.0.0rc14
azure-mgmt-dns==2.1.0
azure-mgmt-keyvault==1.1.0
azure-mgmt-loganalytics==0.2.0
azure-mgmt-managementgroups==0.1.0
azure-mgmt-monitor==0.5.2
azure-mgmt-network==2.7.0
azure-mgmt-rdbms==1.9.0
azure-mgmt-redis==5.0.0
azure-mgmt-resource==3.1.0
azure-mgmt-storage==2.0.0
azure-mgmt-web==0.47.0
msrest==0.6.17
msrestazure==0.6.4
